audio: Fix logging in StreamBluetooth
When logging an error about bits per sample mismatch,
coerce `int8_t` to `int` because the former is handled
the same way as `char` type yielding non-printable
characters.

Bug: 463287369
Test: observe the log

+++ audio/aidl/default/bluetooth/StreamBluetooth.cpp
@@ -218,8 +218,8 @@ bool StreamBluetooth::checkConfigParams(const PcmConfiguration& pcmConfig,
     const int8_t bitsPerSample =
             aidl::android::hardware::audio::common::getPcmSampleSizeInBytes(config.format.pcm) * 8;
     if (bitsPerSample != pcmConfig.bitsPerSample) {
-        LOG(ERROR) << __func__ << ": bits per sample mismatch, stream value=" << bitsPerSample
-                   << ", BT HAL value=" << pcmConfig.bitsPerSample;
+        LOG(ERROR) << __func__ << ": bits per sample mismatch, stream value=" << (int)bitsPerSample
+                   << ", BT HAL value=" << (int)pcmConfig.bitsPerSample;
         return false;
     }
     return true;
